Eliminate "comparison between signed and unsigned integer" warnings.

+++ source/Commands/CommandObjectFrame.cpp
@@ -201,14 +201,14 @@ public:
                 
                 if (m_options.relative_frame_offset < 0)
                 {
-                    if (frame_idx >= -m_options.relative_frame_offset)
+                    if (frame_idx >= (uint32_t)-m_options.relative_frame_offset)
                         frame_idx += m_options.relative_frame_offset;
                     else
                         frame_idx = 0;
                 }
                 else if (m_options.relative_frame_offset > 0)
                 {
-                    if (num_frames - frame_idx > m_options.relative_frame_offset)
+                    if (num_frames - frame_idx > (uint32_t)m_options.relative_frame_offset)
                         frame_idx += m_options.relative_frame_offset;
                     else
                         frame_idx = num_frames - 1;
